fix: gnark-crypto version

+++ base/gfspclient/metadata.go
@@ -51,7 +51,7 @@ func (s *GfSpClient) ListDeletedObjectsByBlockNumberRange(ctx context.Context, s
 		log.CtxErrorw(ctx, "failed to list deleted objects by block number range", "error", err)
 		return nil, uint64(0), ErrRPCUnknownWithDetail("failed to list deleted objects by block number range, error: ", err)
 	}
-	return resp.GetObjects(), uint64(resp.GetEndBlockNumber()), nil
+	return resp.GetObjects(), resp.GetEndBlockNumber(), nil
 }
 
 func (s *GfSpClient) GetUserBuckets(ctx context.Context, account string, includeRemoved bool, opts ...grpc.DialOption) ([]*types.VGFInfoBucket, error) {
